wip: refactor: Conflict chats
