+++ src/entities/chat/mock-conflict.ts
@@ -2,6 +2,7 @@ import {
   chatTypes,
   ConflictChatInfo,
   ConflictChatsTupleMetaInterface,
+  MessageInterface,
   RecipientConflictChatMetaInterface,
   TaskChatInfo,
   TaskChatMetaInterface,
@@ -9,10 +10,6 @@ import {
 } from 'shared/types/chat.types';
 import { mockChatMessages } from './mock-messages';
 import { mockAdmin, mockRecipient, mockVolunteer } from './mock-users';
-import {
-  mockConflictRecipientMessages,
-  mockConflictVolunteerMessages,
-} from './mock-conflict-messages';
 
 /* #####################
 ################### USER
@@ -35,6 +32,47 @@ export const mockConflictTasks: TaskChatInfo = {
   meta: mockConflictTaskChatMeta,
   chats: mockChatMessages,
 };
+
+/* #####################
+############### MESSAGES
+##################### */
+export const mockConflictVolunteerMessages: MessageInterface[] = [
+  {
+    _id: '1',
+    author: mockVolunteer,
+    attaches: [],
+    chatId: '112',
+    body: 'Он не пришёл!',
+    createdAt: new Date(2024, 10, 15, 0, 0, 0),
+  },
+  {
+    _id: '1',
+    author: mockVolunteer,
+    attaches: [],
+    chatId: '112',
+    body: 'Я всё облазил!',
+    createdAt: new Date(2024, 10, 16, 0, 0, 0),
+  },
+];
+
+export const mockConflictRecipientMessages: MessageInterface[] = [
+  {
+    _id: '1',
+    author: mockRecipient,
+    attaches: [],
+    chatId: '911',
+    body: 'Волонтёр не пришёл!',
+    createdAt: new Date(2024, 10, 15, 0, 0, 0),
+  },
+  {
+    _id: '2',
+    author: mockRecipient,
+    attaches: [],
+    chatId: '911',
+    body: 'Я замерз.',
+    createdAt: new Date(2024, 10, 16, 5, 0, 0),
+  },
+];
 
 /* #####################
 ################## ADMIN
